Refactor a code
