+++ app/build.gradle
@@ -44,26 +44,27 @@ android {
 
 dependencies {
     kapt "com.github.bumptech.glide:compiler:4.10.0"
-    kapt "com.google.dagger:hilt-android-compiler:2.31.2-alpha"
+    kapt "com.google.dagger:hilt-android-compiler:2.33-beta"
     implementation "androidx.appcompat:appcompat:1.2.0"
     implementation "androidx.constraintlayout:constraintlayout:2.0.4"
     implementation "androidx.core:core-ktx:1.3.2"
-    implementation "androidx.fragment:fragment-ktx:1.3.2"
+    implementation "androidx.fragment:fragment-ktx:1.3.3"
     implementation "androidx.lifecycle:lifecycle-extensions:2.2.0"
     implementation "androidx.lifecycle:lifecycle-livedata-ktx:2.3.1"
     implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:2.3.1"
-    implementation "androidx.navigation:navigation-fragment-ktx:2.3.4"
-    implementation "androidx.navigation:navigation-ui-ktx:2.3.4"
-    implementation "androidx.recyclerview:recyclerview:1.1.0"
-    implementation "com.github.bumptech.glide:glide:4.10.0"
+    implementation "androidx.navigation:navigation-fragment-ktx:2.3.5"
+    implementation "androidx.navigation:navigation-ui-ktx:2.3.5"
+    implementation "androidx.recyclerview:recyclerview:1.2.0"
+    implementation "com.github.bumptech.glide:glide:4.11.0"
     implementation "com.google.android.material:material:1.3.0"
     implementation "com.google.code.gson:gson:2.8.6"
     implementation "com.squareup.retrofit2:converter-gson:2.9.0"
     implementation "com.squareup.retrofit2:retrofit:2.9.0"
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:1.4.2"
-    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:1.4.2"
+    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:1.4.3"
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
-    implementation "com.google.dagger:hilt-android:2.31.2-alpha"
+    implementation "com.google.dagger:hilt-android:2.33-beta"
+    implementation 'com.squareup.okhttp3:logging-interceptor:3.11.0'
 
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.2'
