+++ app/src/main/java/com/depromeet/muyaho/api/ApiHelper.kt
@@ -1,0 +1,4 @@
+package com.depromeet.muyaho.api
+
+interface ApiHelper {
+}
