+++ app/src/main/java/com/depromeet/muyaho/api/ApiHelperImpl.kt
@@ -1,0 +1,4 @@
+package com.depromeet.muyaho.api
+
+class ApiHelperImpl {
+}
