+++ app/src/main/java/com/depromeet/muyaho/api/ApiService.kt
@@ -1,0 +1,9 @@
+package com.depromeet.muyaho.api
+
+import retrofit2.Response
+import retrofit2.http.GET
+
+interface ApiService {
+    @GET("employees")
+    suspend fun getEmployees(): Response<Unit>
+}
