+++ app/src/main/java/com/depromeet/muyaho/ui/MainActivity.kt
@@ -1,7 +1,8 @@
-package com.depromeet.muyaho
+package com.depromeet.muyaho.ui
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import com.depromeet.muyaho.R
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
